+++ src/main/java/com/mvc/coinsimulation/dto/common/OrderBookUnit.java
@@ -4,12 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class OrderBookUnit {
-    private Double askPrice;
-    private Double bidPrice;
-    private Double askSize;
-    private Double bidSize;
+    private BigDecimal askPrice;
+    private BigDecimal bidPrice;
+    private BigDecimal askSize;
+    private BigDecimal bidSize;
 }
