+++ src/main/java/com/mvc/coinsimulation/dto/response/AssetResponse.java
@@ -3,11 +3,13 @@ package com.mvc.coinsimulation.dto.response;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Builder
 @Data
 public class AssetResponse {
     private String code;
-    private Double amount;
-    private Double buyingPrice;
-    private Double currentPrice;
+    private BigDecimal amount;
+    private BigDecimal buyingPrice;
+    private BigDecimal currentPrice;
 }
