+++ src/main/java/com/mvc/coinsimulation/dto/response/ExecutionResponse.java
@@ -4,16 +4,17 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import com.mvc.coinsimulation.enums.Gubun;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
 public class ExecutionResponse {
     private Long id;
     private Gubun gubun;
-    private Double amount;
+    private BigDecimal amount;
     private String code;
-    private Double price;
-    private Double totalPrice;
+    private BigDecimal price;
+    private BigDecimal totalPrice;
     @JsonFormat(pattern = "yyyy-MM-dd kk:mm:ss")
     private LocalDateTime dateTime;
 }
