+++ src/main/java/com/mvc/coinsimulation/dto/response/OrderResponse.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import com.mvc.coinsimulation.enums.Gubun;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -11,8 +12,8 @@ public class OrderResponse {
     private Long id;
     private String code;
     private Gubun gubun;
-    private Double price;
-    private Double amount;
+    private BigDecimal price;
+    private BigDecimal amount;
     @JsonFormat(pattern = "yyyy-MM-dd kk:mm:ss")
     private LocalDateTime dateTime;
 }
