+++ src/main/java/com/mvc/coinsimulation/dto/response/UserResponse.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -12,5 +14,5 @@ import lombok.NoArgsConstructor;
 public class UserResponse {
     private String nickname;
     private String profile;
-    private Double cash;
+    private BigDecimal cash;
 }
